refactor: view endpoint to hexa
